backend/requirements-test.txt
# Testing dependencies for Smart Home Energy Services

# Core testing framework
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0

# FastAPI testing
fastapi[testing]>=0.100.0
httpx>=0.24.0

# Database and ORM
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# Authentication and security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
email-validator>=2.1.0

# Data processing and AI
pandas>=2.0.0
numpy>=1.24.0
textblob>=0.17.0
nltk>=3.8.0
spacy>=3.7.0
openai>=1.3.0
tiktoken>=0.5.0

# Utilities
python-multipart>=0.0.6
python-dotenv>=1.0.0
redis>=4.5.0
requests>=2.31.0